LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-mutation-insertrule-expected.txt


FAIL Fetched sheet replaces placeholder including manually-inserted rules. assert_equals: Before insertRule: expected 1 placeholder(s). expected 1 but got 0
